code splitting, more visual improvements
